% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coefficients.R
\name{compute_coefficients}
\alias{compute_coefficients}
\title{Compute the simple coefficients for several models}
\usage{
compute_coefficients(
  sample = NULL,
  md_imputation = "multiple",
  md_method = "pmm",
  size = NA,
  formulas = NULL,
  est_method = "felm",
  family = NA
)
}
\arguments{
\item{sample}{Tibble. Initial sample.}

\item{md_imputation}{Character. Kind of imputation ("multiple","single","deletion")}

\item{md_method}{Character. Method for the imputation of missing data ("pmm", "mean", "sample", "deletion").}

\item{size}{Integer. Size of the sample for each draw. If NA, use the sample size.}

\item{formulas}{Vector of formulas. The various models to run on a sample.}

\item{est_method}{Character. Method for the regression ("glm", "felm", or "sysfit").}

\item{family}{Character. Family for glm (e.g. "gaussian", "binomial") .}
}
\value{
A tibble gwith all the coefficients associating each dependent, independent, and moderating variables.
}
\description{
Compute the simple coefficients for several models
}
\seealso{
estimate
}
