% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_path.R
\name{direct_path}
\alias{direct_path}
\title{Compute the moderated direct paths coefficients for analayses of moderated mediations.}
\usage{
direct_path(coefficients, moderations)
}
\arguments{
\item{coefficients}{Tibble. Output of the function compute_coefficients.}

\item{moderations}{Tibble. Output of the function retrieve_values.}
}
\value{
A tibble gathering direct and total moderated path coefficients.
}
\description{
Compute the moderated direct paths coefficients for analayses of moderated mediations.
}
\seealso{
analyze_path
}
