% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{formula_expr}
\alias{formula_expr}
\title{Parse Formula Expressions}
\usage{
formula_expr(formula)
}
\arguments{
\item{formula}{a \code{formula} object of the form \code{y ~ x | g}.}
}
\value{
Returns a list of named character strings:
\tabular{ll}{
  \code{response}  \tab :  left-hand side (lhs) of \code{formula}.
    Everything to the left of the \code{~} operator. \cr
  \code{model}     \tab :  right-hand side (rhs) of \code{formula}.
    Everything to the right of the \code{~} operator. \cr
  \code{covariate} \tab :  covariate expression from the rhs.
    Everything between the \code{~} and \code{|} operators. \cr
  \code{condition} \tab :  conditioning expression from the rhs.
    Everything to the right of the \code{|} operator.
}
}
\description{
Extracts component expressions of a \code{formula} object.
Returns a list containing the response (left-hand side),
model (right-hand side), covariates (model without conditioning
expressions), and conditioning (grouping) expressions.
}
\examples{
fe <- formula_expr(y ~ x | g)
fe$response      # "y"
fe$model         # "x | g"
fe$covariate     # "x"
fe$condition     # "g"

f <- "y1 + y2 ~ x1 + x2 | g1/g2"
fe <- formula_expr(as.formula(f))
fe$response      # "y1 + y2"
fe$model         # "x1 + x2 | g1/g2"
fe$covariate     # "x1 + x2"
fe$condition     # "g1/g2"
}
\author{
Nathan Lucas
}
