% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{formula_lhs}
\alias{formula_lhs}
\title{Formula Left-Hand Side}
\usage{
formula_lhs(x)
}
\arguments{
\item{x}{a formula object of the form \code{y ~ x | g}.}
}
\value{
Returns a character string containing the response
  expression of \code{formula}.
}
\description{
Extracts the response expression of \code{formula}
(everything to the left of the \code{~} operator).
}
\examples{
formula_lhs(y ~ x | g)                   # "y"
formula_lhs(y1 + y2 ~ x1 + x2 | g1/g2)   # "y1 + y2"
}
\author{
Nathan Lucas
}
