% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{formula_rhs}
\alias{formula_rhs}
\title{Formula Right-Hand Side}
\usage{
formula_rhs(x)
}
\arguments{
\item{x}{a formula object of the form \code{y ~ x | g}.}
}
\value{
Returns a character string containing the model term
  expression of \code{formula}.
}
\description{
Extracts model term expression of \code{formula}
(everything to the right of the \code{~} operator).
}
\examples{
formula_rhs(y ~ x | g)                   # "x | g"
formula_rhs(y1 + y2 ~ x1 + x2 | g1/g2)   # "x1 + x2 | g1/g2"
}
\author{
Nathan Lucas
}
