% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect_path.R
\name{indirect_path}
\alias{indirect_path}
\title{Compute the moderated indirect paths coefficients for analayses of moderated mediations (limited to one medietor).}
\usage{
indirect_path(coefficients, moderations, mediation)
}
\arguments{
\item{coefficients}{Tibble. Output of the function compute_coefficients.}

\item{moderations}{Tibble. Output of the function retrieve_values.}

\item{mediation}{List. List of vectors specifying the paths, from the first independent variable to the last dependent variable through the moderators in order.}
}
\value{
A tibble gathering indirect and total moderated path coefficients.
}
\description{
Compute the moderated indirect paths coefficients for analayses of moderated mediations (limited to one medietor).
}
\seealso{
analyze_path
}
