% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_edwards2007.R
\docType{data}
\name{mm_edwards2007}
\alias{mm_edwards2007}
\title{Data from Edwards and Lambert 2007 for tests of moderated mediation.}
\format{A tiblle with 1,307 observations and 17 variables:
\itemize{
\item gen. Gender (raw)
\item cen. Centrality (raw)
\item fbk. Feedback (raw)
\item sat. Satisfaction (raw)
\item com. Commitment (raw)
\item cenc. Centrality (scaled)
\item fbkc. Feedback (scaled)
\item satc. Satisfaction (scaled)
\item comc. Commitment (scaled)
\item fbkgen. Interaction between Feedback and Gender
\item satgen. Interaction between Satisfaction and Gender
\item fbkcen. Interaction between Feedback and Centrality
\item satcen. Interaction between Satisfaction and Centrality
\item fbkcgen. Interaction between Feedback and Gender
\item satcgen. Interaction between Satisfaction and Gender
\item fbkccenc. Interaction between Feedback and Centrality
\item satccenc. Interaction between and Centrality
}}
\source{
\url{http//:dx.doi.org/10.1037/1082-989X.12.1.1.supp}
}
\usage{
data("mm_edwards2007")
}
\description{
A dataset to test the impact of feedback on commitment through satisfaction, moderated by gender or centrality.
}
\references{
{Edwards, J. R., & Lambert, L. S. (2007). Methods for integrating moderation and mediation: a general analytical framework using moderated path analysis. Psychological Methods, 12(1), 1–22.}
}
\keyword{datasets}
