% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlr.R
\docType{package}
\name{modlr}
\alias{modlr}
\alias{modlr-package}
\title{Model fitting utilities for R}
\description{
Utility functions to make model fitting easier and more consistent.
}
\author{
Nathan Lucas
}
