% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_model.R
\name{prep_model}
\alias{prep_model}
\title{Create a formula based on lists of variables. All independent variables will be interacted with the moderator if the latter is specified. Controls are not.}
\usage{
prep_model(
  dependent = NULL,
  independent = c(NA),
  moderator = c(NA),
  control = c(NA),
  fixed_effect = c(NA),
  instrument = c(NA),
  cluster = c(NA)
)
}
\arguments{
\item{dependent}{Character string. Dependent variable.}

\item{independent}{Character string. Variables of interest.}

\item{moderator}{Character vector. Moderating variable.}

\item{control}{Character vector. Control variables.}

\item{fixed_effect}{Character vector. Fixed effects.}

\item{instrument}{Character vector. Instruments.}

\item{cluster}{Character vector. Clusterings for errors.}
}
\value{
A formula to run in regressions. Interactions are introduced under the form "independent_x_moderator" and therefore the corresponding variable must be created in the dataset.
}
\description{
Create a formula based on lists of variables. All independent variables will be interacted with the moderator if the latter is specified. Controls are not.
}
\examples{
library(modlR)
model <- prep_model(dependent = "turnover",
                    independent = c("distributive", "procedural"),
                    moderator = "pay",
                    control = c("age","gender"),
                    fixed_effect = c("team","period"),
                    instrument = c("chock","manip"),
                    cluster = c("country"))
model
}
