% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sample.R
\name{prepare_sample}
\alias{prepare_sample}
\title{Compute the paths coefficients for analayses of moderated mediations.}
\usage{
prepare_sample(sample, md_imputation, md_method, size)
}
\arguments{
\item{sample}{Tibble. Table containing the data.}

\item{md_imputation}{Character. Kind of imputation ("multiple","single","deletion")}

\item{md_method}{Character. Method for the imputation of missing data ("pmm", "mean", "sample", "deletion").}

\item{size}{Integer. Size of the sample for each draw. If NA, use the sample size.}
}
\value{
A tibble gathering the regression coefficients for each draw.
}
\description{
Compute the paths coefficients for analayses of moderated mediations.
}
\references{
Arne Henningsen and Jeff D. Hamann (2007). systemfit: A Package for Estimating Systems of Simultaneous Equations in R. Journal of Statistical Software 23(4), 1-40. http://www.jstatsoft.org/v23/i04/.
}
