% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_modval.R
\name{retrieve_modval}
\alias{retrieve_modval}
\title{Retrieve values of the moderators necessary for analyses of moderated mediations.}
\usage{
retrieve_modval(sample, moderators, types, allval)
}
\arguments{
\item{sample}{Tibble. Table containing the data.}

\item{moderators}{Character vector. vector speciying which variables moderate the relationships.}

\item{types}{Character vector. Whether the corresponding moderator is "discrete" (actually binary) or "continuous".}

\item{allval}{Logical. Whether all values of the moderating variables should be retrieved (TRUE) or just two representative values.}
}
\value{
A tibble gathering the requested values for the moderators.
}
\description{
Retrieve values of the moderators necessary for analyses of moderated mediations.
}
\seealso{
analyze_path
}
